requirements.txt
torch>=2.0.0
transformers[torch]>=4.30.0
accelerate>=0.26.0
datasets>=2.12.0
pandas>=2.0.0
numpy>=1.24.0
gradio>=3.35.2
python-dotenv>=1.0.0
loguru>=0.7.0
pytest>=7.3.1
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
